apps/evals/pyproject.toml
[project]
name = "crescendai-evals"
version = "0.1.0"
description = "Evaluation harness for CrescendAI (pipeline, model, memory)"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "groq>=0.4.0",
    "requests>=2.31.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
model = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
]
memory = [
    "sentence-transformers>=2.0.0",
    "pytest>=7.0.0",
]
all = [
    "crescendai-evals[model,memory]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "pipeline", "model", "memory", "inference"]

[tool.pytest.ini_options]
markers = ["integration: marks integration tests (require live API)"]
